requirements.txt
# Specify where to install requirements from.
--index-url https://pypi.python.org/simple/

# Needed packages/versions.
win32com.client
pythoncom